Update README.md
The rhcos-X.X.X-x86_64-installer.x86_64.iso is replaced with rhcos-X.X.X-x86_64-live.x86_64.iso 
Added info on creating OCP network in case of using VirtualBox or ProxMox (I tried it myself on both VirtualBox and ProxMox) 
Fixed a small mistake in the third file to be copied to the helper node (although it was correct in the youtube video)
Added tcp protocol for DNS port 53 in helper node's internal zone. Without it the cluster won't resolve internet addresses like quay.io or github.com for sample apps. This only happened with the new version 4.9.
Added an optional installation command for the CoreOS after boot. It's a little bit shorter.